[fix] adjust ui

+++ app/src/main/java/com/jie/calculator/calculator/util/SystemUtil.java
@@ -34,7 +34,7 @@ public class SystemUtil {
         return statusHeight;
     }
 
-    public static int dp2px(Context context, int dp) {
+    public static int dp2px(Context context, float dp) {
         return (int) TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, dp, context.getResources().getDisplayMetrics());
     }
 }
